+++ tests/TransformersCompat.hs
@@ -1,18 +1,16 @@
 {-# LANGUAGE CPP                        #-}
--- {-# LANGUAGE DeriveDataTypeable         #-}
+{-# LANGUAGE DeriveDataTypeable         #-}
 {-# LANGUAGE DeriveFoldable             #-}
 {-# LANGUAGE DeriveFunctor              #-}
 {-# LANGUAGE DeriveTraversable          #-}
 {-# LANGUAGE FlexibleContexts           #-}
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
-{-# LANGUAGE StandaloneDeriving         #-}
 
 #if __GLASGOW_HASKELL__ >= 702
 {-# LANGUAGE DeriveGeneric              #-}
-#endif
-
-#if __GLASGOW_HASKELL__ >= 706
-{-# LANGUAGE PolyKinds                  #-}
+#else
+{-# LANGUAGE TemplateHaskell            #-}
+{-# LANGUAGE TypeFamilies               #-}
 #endif
 
 {-|
@@ -36,26 +34,27 @@ module TransformersCompat (
     showsUnaryWith,
     showsBinaryWith,
     -- * Conversion between @Text-@ and @String@ @Show1@/@Show2@
-    FromStringShow1(..),
-    FromStringShow2(..)
+    FromStringShow1(..), FromTextShow1(..),
+    FromStringShow2(..), FromTextShow2(..)
     ) where
 
-#include "generic.h"
 #include "inline.h"
 
 import           Control.Applicative (Const(..))
 
--- #if __GLASGOW_HASKELL__ >= 708
--- import           Data.Data (Data, Typeable)
--- #endif
+import           Data.Bifunctor (Bifunctor(..))
+#if __GLASGOW_HASKELL__ >= 708
+import           Data.Data (Data, Typeable)
+#endif
 import           Data.Functor.Identity (Identity(..))
-import           Data.Text.Lazy.Builder (Builder)
 
 #if __GLASGOW_HASKELL__ >= 702
 import           GHC.Generics (Generic)
-# if defined(__LANGUAGE_DERIVE_GENERIC1__)
+# if __GLASGOW_HASKELL__ >= 706
 import           GHC.Generics (Generic1)
 # endif
+#else
+import qualified Generics.Deriving.TH as Generics
 #endif
 
 import           Prelude ()
@@ -64,7 +63,8 @@ import           Prelude.Compat
 import           Text.Read (Read(..), readListPrecDefault)
 
 import           TextShow (TextShow(showbPrec), TextShow1(..), TextShow2(..),
-                           fromString, toString)
+                           FromStringShow(..), FromTextShow(..),
+                           showsToShowb, showbToShows, showbPrec1, showbPrec2)
 
 -- | Lifting of the 'Show' class to unary type constructors.
 class Show1 f where
@@ -75,34 +75,33 @@ class Show1 f where
 showsPrec1 :: (Show1 f, Show a) => Int -> f a -> ShowS
 showsPrec1 = showsPrecWith showsPrec
 
+-- | The 'TextShow1' instance for 'FromStringShow1' is based on its @String@
+-- 'Show1' instance. That is,
+--
+-- @
+-- showbPrecWith sp p ('FromStringShow1' x) =
+--     'showsToShowb' ('showsPrecWith' ('showbToShows' sp)) p x
+-- @
+--
+-- /Since: ?.?/
 newtype FromStringShow1 f a = FromStringShow1 { fromStringShow1 :: f a }
   deriving ( Eq
+           , Functor
+           , Foldable
 #if __GLASGOW_HASKELL__ >= 702
            , Generic
-# if defined(__LANGUAGE_DERIVE_GENERIC1__)
+# if __GLASGOW_HASKELL__ >= 706
            , Generic1
 # endif
 #endif
            , Ord
+           , Show1
+           , Traversable
+#if __GLASGOW_HASKELL__ >= 708
+           , Data
+           , Typeable
+#endif
            )
-
-deriving instance Show1       f => Show1       (FromStringShow1 f)
-deriving instance Functor     f => Functor     (FromStringShow1 f)
-deriving instance Foldable    f => Foldable    (FromStringShow1 f)
-deriving instance Traversable f => Traversable (FromStringShow1 f)
--- #if __GLASGOW_HASKELL__ >= 708
--- deriving instance Typeable FromStringShow1
--- deriving instance (Data (f a), Typeable f, Typeable a) =>
---     Data (FromStringShow1 f a)
--- #endif
-
--- | Convert a @Builder@-based show function to a @ShowS@-based one.
-showbToShows :: (Int -> a -> Builder) -> Int -> a -> ShowS
-showbToShows sp p = showString . toString . sp p
-
--- | Convert a @ShowS@-based show function to a @Builder@-based one.
-showsToShowb :: (Int -> a -> ShowS) -> Int -> a -> Builder
-showsToShowb sp p x = fromString $ sp p x ""
 
 instance Read (f a) => Read (FromStringShow1 f a) where
     readPrec = FromStringShow1 <$> readPrec
@@ -116,8 +115,8 @@ instance (Show1 f, Show a) => TextShow (FromStringShow1 f a) where
     INLINE_INST_FUN(showbPrec)
 
 instance Show1 f => TextShow1 (FromStringShow1 f) where
-    showbPrecWith sp p (FromStringShow1 x) =
-        fromString $ showsPrecWith (showbToShows sp) p x ""
+    showbPrecWith sp p =
+        showsToShowb (showsPrecWith $ showbToShows sp) p . fromStringShow1
     INLINE_INST_FUN(showbPrecWith)
 
 instance (Show1 f, Show a) => Show (FromStringShow1 f a) where
@@ -134,8 +133,67 @@ class Show2 f where
 showsPrec2 :: (Show2 f, Show a, Show b) => Int -> f a b -> ShowS
 showsPrec2 = showsPrecWith2 showsPrec showsPrec
 
-newtype FromStringShow2 f a b = FromStringShow2 (f a b)
+-- | The @String@ 'Show1' instance for 'FromTextShow1' is based on its
+-- 'TextShow1' instance. That is,
+--
+-- @
+-- showsPrecWith sp p ('FromTextShow1' x) =
+--     'showbToShows' ('showbPrecWith' ('showsToShowb' sp)) p x
+-- @
+--
+-- /Since: ?.?/
+newtype FromTextShow1 f a = FromTextShow1 { fromTextShow1 :: f a }
   deriving ( Eq
+           , Functor
+           , Foldable
+#if __GLASGOW_HASKELL__ >= 702
+           , Generic
+# if __GLASGOW_HASKELL__ >= 706
+           , Generic1
+# endif
+#endif
+           , Ord
+           , TextShow1
+           , Traversable
+#if __GLASGOW_HASKELL__ >= 708
+           , Data
+           , Typeable
+#endif
+           )
+
+instance Read (f a) => Read (FromTextShow1 f a) where
+    readPrec = FromTextShow1 <$> readPrec
+    INLINE_INST_FUN(readPrec)
+
+    readListPrec = readListPrecDefault
+    INLINE_INST_FUN(readListPrec)
+
+instance (TextShow1 f, TextShow a) => Show (FromTextShow1 f a) where
+    showsPrec = showsPrecWith (showbToShows showbPrec)
+    INLINE_INST_FUN(showsPrec)
+
+instance TextShow1 f => Show1 (FromTextShow1 f) where
+    showsPrecWith sp p =
+        showbToShows (showbPrecWith $ showsToShowb sp) p . fromTextShow1
+    INLINE_INST_FUN(showsPrecWith)
+
+instance (TextShow1 f, TextShow a) => TextShow (FromTextShow1 f a) where
+    showbPrec = showbPrec1
+    INLINE_INST_FUN(showbPrec)
+
+-- | The 'TextShow2' instance for 'FromStringShow2' is based on its @String@
+-- 'Show2' instance. That is,
+--
+-- @
+-- showbPrecWith2 sp1 sp2 p ('FromStringShow2' x) =
+--     'showsToShowb' ('showsPrecWith2' ('showbToShows' sp1) ('showbToShows' sp2)) p x
+-- @
+--
+-- /Since: ?.?/
+newtype FromStringShow2 f a b = FromStringShow2 { fromStringShow2 :: f a b }
+  deriving ( Eq
+           , Functor
+           , Foldable
 #if __GLASGOW_HASKELL__ >= 702
            , Generic
 # if defined(__LANGUAGE_DERIVE_GENERIC1__)
@@ -143,17 +201,17 @@ newtype FromStringShow2 f a b = FromStringShow2 (f a b)
 # endif
 #endif
            , Ord
+           , Show2
+           , Traversable
+#if __GLASGOW_HASKELL__ >= 708
+           , Data
+           , Typeable
+#endif
            )
 
-deriving instance Show2       f     => Show2       (FromStringShow2 f)
-deriving instance Functor     (f a) => Functor     (FromStringShow2 f a)
-deriving instance Foldable    (f a) => Foldable    (FromStringShow2 f a)
-deriving instance Traversable (f a) => Traversable (FromStringShow2 f a)
--- #if __GLASGOW_HASKELL__ >= 708
--- deriving instance Typeable FromStringShow2
--- deriving instance (Data (f a b), Typeable f, Typeable a, Typeable b) =>
---     Data (FromStringShow2 f a b)
--- #endif
+instance Bifunctor f => Bifunctor (FromStringShow2 f) where
+    bimap f g = FromStringShow2 . bimap f g . fromStringShow2
+    INLINE_INST_FUN(bimap)
 
 instance Read (f a b) => Read (FromStringShow2 f a b) where
     readPrec = FromStringShow2 <$> readPrec
@@ -171,8 +229,8 @@ instance (Show2 f, Show a) => TextShow1 (FromStringShow2 f a) where
     INLINE_INST_FUN(showbPrecWith)
 
 instance Show2 f => TextShow2 (FromStringShow2 f) where
-    showbPrecWith2 sp1 sp2 p (FromStringShow2 x) =
-        fromString $ showsPrecWith2 (showbToShows sp1) (showbToShows sp2) p x ""
+    showbPrecWith2 sp1 sp2 p =
+        showsToShowb (showsPrecWith2 (showbToShows sp1) (showbToShows sp2)) p . fromStringShow2
     INLINE_INST_FUN(showbPrecWith2)
 
 instance (Show2 f, Show a, Show b) => Show (FromStringShow2 f a b) where
@@ -182,6 +240,66 @@ instance (Show2 f, Show a, Show b) => Show (FromStringShow2 f a b) where
 instance (Show2 f, Show a) => Show1 (FromStringShow2 f a) where
     showsPrecWith = showsPrecWith2 showsPrec
     INLINE_INST_FUN(showsPrecWith)
+
+-- | The @String@ 'Show2' instance for 'FromTextShow2' is based on its
+-- 'TextShow2' instance. That is,
+--
+-- @
+-- showsPrecWith2 sp1 sp2 p ('FromTextShow2' x) =
+--     'showbToShows' ('showbPrecWith2' ('showsToShowb' sp1) ('showsToShowb' sp2)) p x
+-- @
+--
+-- /Since: ?.?/
+newtype FromTextShow2 f a b = FromTextShow2 { fromTextShow2 :: f a b }
+  deriving ( Eq
+           , Functor
+           , Foldable
+#if __GLASGOW_HASKELL__ >= 702
+           , Generic
+# if defined(__LANGUAGE_DERIVE_GENERIC1__)
+           , Generic1
+# endif
+#endif
+           , Ord
+           , TextShow2
+           , Traversable
+#if __GLASGOW_HASKELL__ >= 708
+           , Data
+           , Typeable
+#endif
+           )
+
+instance Bifunctor f => Bifunctor (FromTextShow2 f) where
+    bimap f g = FromTextShow2 . bimap f g . fromTextShow2
+    INLINE_INST_FUN(bimap)
+
+instance Read (f a b) => Read (FromTextShow2 f a b) where
+    readPrec = FromTextShow2 <$> readPrec
+    INLINE_INST_FUN(readPrec)
+
+    readListPrec = readListPrecDefault
+    INLINE_INST_FUN(readListPrec)
+
+instance (TextShow2 f, TextShow a, TextShow b) => Show (FromTextShow2 f a b) where
+    showsPrec = showsPrecWith (showbToShows showbPrec)
+    INLINE_INST_FUN(showsPrec)
+
+instance (TextShow2 f, TextShow a) => Show1 (FromTextShow2 f a) where
+    showsPrecWith = showsPrecWith2 (showbToShows showbPrec)
+    INLINE_INST_FUN(showsPrecWith)
+
+instance TextShow2 f => Show2 (FromTextShow2 f) where
+    showsPrecWith2 sp1 sp2 p =
+        showbToShows (showbPrecWith2 (showsToShowb sp1) (showsToShowb sp2)) p . fromTextShow2
+    INLINE_INST_FUN(showsPrecWith2)
+
+instance (TextShow2 f, TextShow a, TextShow b) => TextShow (FromTextShow2 f a b) where
+    showbPrec = showbPrec2
+    INLINE_INST_FUN(showbPrec)
+
+instance (TextShow2 f, TextShow a) => TextShow1 (FromTextShow2 f a) where
+    showbPrecWith = showbPrecWith2 showbPrec
+    INLINE_INST_FUN(showbPrecWith)
 
 -------------------------------------------------------------------------------
 
@@ -243,3 +361,20 @@ instance (Show a, Show b, Show c) => Show2 ((,,,,) a b c) where
                      . shows c . showChar ','
                      . sp1 0 d . showChar ','
                      . sp2 0 e . showChar ')'
+
+-- TODO: Move these instance into text-show itself once transformers is updated
+instance Show1 FromStringShow where
+    showsPrecWith sp p = sp p . fromStringShow
+    INLINE_INST_FUN(showsPrecWith)
+
+instance Show1 FromTextShow where
+    showsPrecWith sp p =
+        showbToShows (showsToShowb sp) p . fromTextShow
+    INLINE_INST_FUN(showsPrecWith)
+
+-------------------------------------------------------------------------------
+
+#if __GLASGOW_HASKELL__ < 702
+$(Generics.deriveAll ''FromStringShow1)
+$(Generics.deriveAll ''FromStringShow2)
+#endif